animate-studio/requirements.txt
# ═══════════════════════════════════════════════════════════
# AniMate Studio — Dependencies
# ═══════════════════════════════════════════════════════════
# Install: pip install -r requirements.txt
# Requires: Python 3.11+, CUDA 12.1+, FFmpeg in PATH

# ── Core Framework ───────────────────────────────────────
torch>=2.4.0
torchvision>=0.19.0
torchaudio>=2.4.0

# ── Diffusion / AI Models ───────────────────────────────
diffusers>=0.30.0
transformers>=4.44.0
accelerate>=0.33.0
safetensors>=0.4.4
peft>=0.12.0
tiktoken>=0.7.0
sentencepiece>=0.2.0
protobuf>=4.25.0
# xformers — no Python 3.13 wheel available; pipeline uses default attention

# ── Web UI ───────────────────────────────────────────────
gradio>=4.40.0

# ── API / Networking ────────────────────────────────────
openai>=1.35.0
requests>=2.32.0

# ── Audio ────────────────────────────────────────────────
edge-tts>=6.1.12

# ── Image / Video Processing ────────────────────────────
Pillow>=10.4.0
numpy>=1.26.0
opencv-python>=4.10.0
imageio[ffmpeg]>=2.34.0

# ── AI Upscaling ────────────────────────────────────────
realesrgan>=0.3.0
basicsr>=1.4.2
facexlib>=0.3.0
gfpgan>=1.3.8

# ── Config / Utilities ──────────────────────────────────
pyyaml>=6.0.1
psutil>=5.9.0
